Refactor State
